requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.8
pyqt6>=6.5.0
requests>=2.31.0
dnspython>=2.4.0
email-validator>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
